+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## UNRELEASED
 
+IMPROVEMENTS:
+
+* agent: add option to configure max request length for `/v1/txn` endpoint [[GH-7388](https://github.com/hashicorp/consul/pull/7388)]
+
+BUGFIXES:
+
+* sessions: Fixed backwards incompatibility with 1.6.x and earlier [[GH-7395](https://github.com/hashicorp/consul/issues/7395)][[GH-7398](https://github.com/hashicorp/consul/pull/7398)]
+
 ## 1.7.1 (February 20, 2020)
 
 IMPROVEMENTS: